implementacion de boton para atender citas

+++ src/main/java/com/ubam/dentcare_plus/dto/cliente/ActividadRecienteDTO.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class ActividadRecienteDTO {
     Integer id;
     LocalDate fecha;
-    Float costo;
+    BigDecimal costo;
     String nombre;
     String estatus;
 }
